% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwd.R
\name{dbwd}
\alias{dbwd}
\alias{pbwd}
\alias{qbwd}
\alias{rbwd}
\title{A Bimodal Weibull Distribution}
\usage{
dbwd(x, alpha, beta, sigma)

pbwd(x, alpha, beta, sigma)

qbwd(p, alpha, beta, sigma)

rbwd(n, alpha, beta, sigma)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{beta}{shape2}

\item{sigma}{shape3}
}
\value{
\code{dbwd} gives the density, \code{pbwd} gives the distribution function, \code{qbwd} gives the quantile function and \code{rbwd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a Bimodal Weibull Distribution with parameters shape and scale.
}
\details{
A Bimodal Weibull distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{<p>f(x)=(alpha/(beta*z))*(1+(1-sigma*x)^2)*(x/beta)^(alpha-1)*exp(-(x/beta)^alpha)</p>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dbwd(1,2,3,4)
dbwd(c(1:3),2,3,4)
dbwd(1,c(1:3),3,4)
dbwd(1,2,c(1:4),4)
dbwd(1,2,3,c(1:4))
dbwd(c(1:3),c(1:3),3,4)
pbwd(c(1:3),2,3,4)
pbwd(1,c(1:3),3,4)
pbwd(1,2,c(1:4),4)
pbwd(1,2,3,c(1:4))
qbwd(c(.7),1,2,3)
qbwd(c(.7),2,2,c(3:5))
rbwd(10,1,3,2)
}
\references{
Vila, R. ve Niyazi Çankaya, M., 2022,
\emph{A bimodal Weibull distribution: properties and inference},Journal of Applied Statistics, 49 (12), 3044-3062.
}
