% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dld.R
\name{ddld}
\alias{ddld}
\alias{pdld}
\alias{qdld}
\alias{rdld}
\title{The discrete Lindley distribution}
\usage{
ddld(x, theta)

pdld(x, theta)

qdld(p, theta)

rdld(n, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{ddld} gives the density, \code{pdld} gives the distribution function, \code{qdld} gives the quantile function and \code{rdld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the discrete Lindley distribution with parameters shape and scale.
}
\details{
The discrete Lindley distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{><em>λ</em><sup><em>x</em></sup>/(1−log(<em>λ</em>))</span></p>}{\eqn{T_n^{(\ell)}}}.
}
\examples{
ddld(c(1:5),2)
ddld(c(1:5),-2)
ddld(-1,2)
ddld(2,c(2:4))
ddld(c(1:5),c(2:5))
pdld(2,1)
pdld(1,c(2:4))
qdld(c(.685,.67),.5)
qdld(.993,2)
rdld(10,1)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
