% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pldd.R
\name{dpldd}
\alias{dpldd}
\title{A Power Log Dagum Distribution}
\usage{
dpldd(x, alpha, beta, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{beta}{shape2}

\item{theta}{shape3}
}
\value{
\code{dpldd} gives the density, \code{ppldd} gives the distribution function, \code{qpldd} gives the quantile function and \code{rpldd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a Power Log Dagum distribution with parameters shape and scale.
}
\details{
A Power Log Dagum Distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dpldd(1,2,3,4)
dpldd(c(1:5),2,3,4)
dpldd(c(1:5),c(2:6),3,4)
dpldd(c(1:5),c(2:6),c(3:5),4)
dpldd(c(1:5),c(2:6),c(3:5),c(4:6))
dpldd(1,2,3,c(4:6))
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
