% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmd.R
\name{dtpmd}
\alias{dtpmd}
\title{The Power Muth Distribution}
\usage{
dtpmd(x, beta, alpha)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{beta}{a shape parameter.}

\item{alpha}{a scale parameter.}
}
\value{
\code{dtpmd} gives the density, \code{ptpmd} gives the distribution function, \code{qtpmd} gives the quantile function and \code{rtpmd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Power Muth distribution with parameters shape and scale.
}
\details{
The Power Muth Distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{"(alpha/(beta^alpha))*(x^(alpha-1))*(exp((x/beta)^alpha)-1)*exp(((x/beta)^alpha)-(exp((x/beta)^alpha)-1))"}}{{\eqn{\\ f\left( x\right) =\frac{\alpha }{\beta ^{\alpha }}x^{\alpha -1}\left\{ \exp\left( \left( \frac{x}{\beta }\right) ^{\alpha }-1\right) \right\} \left\{\exp \left( \left( \frac{x}{\beta }\right) ^{\alpha }-\left( \exp \left(\frac{x}{\beta }\right) ^{\alpha }-1\right) \right) \right\}}}}.
}
\note{
the cumulative hazart ... (a) if ...is increasing in x for any ... (b) if ... then there exist an ... so that h(x) is (strictly) decreasing when ... and... (strictly) when ...
}
\examples{
dtpmd(1,2,3)
dtpmd(2,3,4)
dtpmd(c(1:5),2,3)
dtpmd(c(1:5),c(2:6),3)
dtpmd(c(1:5),c(2:6),c(3:6))
}
\references{
Jodra, P., Gomez, H. W., Jimenez-Gamero, M. D., & Alba-Fernandez, M. V. (2017).
\emph{The power Muth distribution} . Mathematical Modelling and Analysis, 22(2), 186-201.
}
