% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R, R/tpmd.R
\name{demd}
\alias{demd}
\alias{pemd}
\alias{qemd}
\alias{remd}
\title{Estimation in Maxwell distribution with randomly censored data}
\usage{
demd(x, theta)

pemd(x, theta)

qemd(p, theta)

remd(n, theta)

demd(x, theta)

pemd(x, theta)

qemd(p, theta)

remd(n, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{demd} gives the density, \code{pemd} gives the distribution function, \code{qemd} gives the quantile function and \code{remd} generates random deviates.

\code{demd} gives the density, \code{pemd} gives the distribution function, \code{qemd} gives the quantile function and \code{remd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for Estimation in Maxwell distribution with parameters shape and scale.

Density, distribution function, quantile function and random generation for Estimation in Maxwell distribution with parameters shape and scale.
}
\details{
Estimation in Maxwell distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{1t<sub>n</sub><sup>(&#952)</sup>}}{\eqn{}}.

Estimation in Maxwell distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{1t<sub>n</sub><sup>(&#952)</sup>}}{\eqn{}}.
}
\examples{
demd(c(1:5),2)
demd(2,c(2:4))
demd(c(1:5),c(2:5))
demd(c(1:5),-2)
demd(-2,c(2:4))
pemd(c(1:5),2)
qemd(c(.7,.8),2)
qemd(.4,5)
remd(100,1)
demd(c(1:5),2)
demd(2,c(2:4))
demd(c(1:5),c(2:5))
demd(c(1:5),-2)
demd(-2,c(2:4))
pemd(c(1:5),2)
qemd(c(.7,.8),2)
qemd(.4,5)
remd(100,1)
}
\references{
Krishna, H., Vivekanand ve Kumar, K., 2015,
\emph{Estimation in Maxwell distribution with randomly censored data}, Journal of statistical computation and simulation, 85 (17), 3560-3578.

Krishna, H., Vivekanand ve Kumar, K., 2015,
\emph{Estimation in Maxwell distribution with randomly censored data}, Journal of statistical computation and simulation, 85 (17), 3560-3578.
}
