% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epkd.R
\name{epkd}
\alias{epkd}
\alias{depkd}
\alias{pepkd}
\alias{qepkd}
\alias{repkd}
\title{Estimation procedures for kumaraswamy distribution parameters under adaptive type-II hybrid progressive censoring}
\usage{
depkd(x, lambda, alpha)

pepkd(x, lambda, alpha)

qepkd(p, lambda, alpha)

repkd(n, lambda, alpha)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{lambda}{a shape1 parameter.}

\item{alpha}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{depkd} gives the density, \code{pepkd} gives the distribution function, \code{qepkd} gives the quantile function and \code{repkd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for Estimation procedures for kumaraswamy distribution with parameters shape and scale.
}
\details{
...Estimation procedures for kumaraswamy distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
depkd(0.1,2,3)
depkd(c(0.1,0.5),2,3)
depkd(c(0.1,0.5),c(2:6),3)
depkd(c(0.1,0.5),c(2:6),c(3:6))
pepkd(0.5,2,3)
pepkd(c(0.1,0.5),2,3)
pepkd(c(0.1,0.5),c(2:6),3)
pepkd(c(0.1,0.5),c(2:6),c(3:6))
pepkd(-1,c(2:6),c(3:6))
qepkd(.8,1,2)
qepkd(.4,2,5)
repkd(10,1,2)
}
\references{
Kohansal, A. ve Bakouch, H. S., 2021,
\emph{Estimation procedures for Kumaraswamy distribution parameters under adaptive type-II hybrid progressive censoring}, Communications in Statistics-Simulation and Computation, 50 (12), 4059-4078.
}
