% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld.R
\name{dgld}
\alias{dgld}
\alias{pgld}
\alias{qgld}
\alias{rgld}
\title{The gamma-Lomax distribution}
\usage{
dgld(x, a, alpha, beta)

pgld(x, a, alpha, beta)

qgld(p, a, alpha, beta)

rgld(n, a, alpha, beta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{a}{a shape parameter.}

\item{alpha}{a shape1 parameter.}

\item{beta}{a shape2 parameter.}
}
\value{
\code{dgld} gives the density, \code{pgld} gives the distribution function, \code{qgld} gives the quantile function and \code{rgld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the gamma-Lomax distribution with parameters shape and scale.
}
\details{
The gamma-Lomax distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dgld(1,2,3,4)
dgld(c(1:5),2,3,4)
dgld(c(1:5),c(2:6),3,4)
dgld(c(1:5),c(2:6),c(3:6),4)
pgld(1,2,3,4)
pgld(c(1:5),2,3,4)
pgld(c(1:5),c(2:6),3,4)
qgld(c(.7,.8),1,2,3)
qgld(c(.7,.8),c(1,2),2,3)
qgld(.8,2,2,3)
rgld(100,1,2,1)
}
\references{
Cordeiro, G. M., Ortega, E. M. ve Popović, B. V., 2015,
\emph{The gamma-Lomax distribution}, Journal of statistical computation and simulation, 85 (2), 305-319.
}
