% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndd.R
\name{dndd}
\alias{dndd}
\alias{pndd}
\alias{qndd}
\alias{rndd}
\title{A new discrete distribution}
\usage{
dndd(x, theta)

pndd(x, theta)

qndd(p, theta)

rndd(n, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dndd} gives the density, \code{pndd} gives the distribution function, \code{qndd} gives the quantile function and \code{rndd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a new discrete distribution with parameters shape and scale.
}
\details{
A new discrete distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dndd(c(1:5),2)
dndd(-2,c(2:4))
dndd(c(1:5),c(2:5))
dndd(1.1,2)
pndd(c(1:5),2)
pndd(2,c(2:4))
pndd(c(1:5),c(2:5))
pndd(-2,c(2:4))
pndd(.2,c(2:4))
qndd(.5,.2)
qndd(.8,1)
rndd(10,.1)
}
\references{
Bakouch, H. S., Jazi, M. A. ve Nadarajah, S., 2014,
\emph{A new discrete distribution, Statistics}, 48 (1), 200-240.
}
