% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nld.R
\name{dnld}
\alias{dnld}
\alias{pnld}
\alias{qnld}
\alias{rnld}
\title{A new lifetime distribution}
\usage{
dnld(x, lambda, beta)

pnld(x, lambda, beta)

qnld(p, lambda, beta)

rnld(n, lambda, beta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{lambda}{a shape parameter.}

\item{beta}{a shape1 parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dnld} gives the density, \code{pnld} gives the distribution function, \code{qnld} gives the quantile function and \code{rnld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a new lifetime distribution with parameters shape and scale.
}
\details{
A new lifetime distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dnld(1,2,3)
dnld(c(1:5),2,3)
dnld(c(1:5),c(2:6),3)
dnld(c(1:5),c(2:6),c(3:6))
pnld(1,2,3)
pnld(c(1:5),2,3)
pnld(c(1:5),c(2:6),3)
pnld(c(1:5),c(2:6),c(3:5))
pnld(-5,c(2:6),3)
qnld(.8,2,3)
rnld(10,1,2)
}
\references{
Kuş, C., 2007,
\emph{A new lifetime distribution}, Computational Statistics & Data Analysis, 51 (9), 4497-4509.
}
