% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noPDD.R
\name{noPDD}
\alias{noPDD}
\alias{dnoPDD}
\alias{pnoPDD}
\alias{qnoPDD}
\alias{rnoPDD}
\title{A new one parameter discrete distribution and its applications}
\usage{
dnoPDD(x, theta)

pnoPDD(x, theta)

qnoPDD(p, theta)

rnoPDD(n, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dnoPDD} gives the density, \code{pnoPDD} gives the distribution function, \code{qnoPDD} gives the quantile function and \code{rnoPDD} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a new one parameter discrete distribution with parameters shape and scale.
}
\details{
A new one parameter discrete distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dnoPDD(c(1:5),2)
dnoPDD(2,c(2:4))
dnoPDD(c(1:5),c(2:5))
dnoPDD(c(1:5),-2)
dnoPDD(-2,c(2:4))
pnoPDD(c(1:5),2)
pnoPDD(2,c(2:4))
pnoPDD(c(1:5),c(2:5))
pnoPDD(-2,c(2:4))
pnoPDD(.2,c(2:4))
qnoPDD(.1,1)
rnoPDD(10,1)
}
\references{
Shukla, K. K., Shanker, R. ve Tiwari, M. K., 2022,
\emph{A new one parameter discrete distribution and its applications}, Journal of Statistics and Management Systems, 25 (1), 269-283.
}
