% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{domd}
\alias{domd}
\alias{pomd}
\alias{qomd}
\alias{romd}
\title{on the muth distribution}
\usage{
domd(x, alpha)

pomd(x, alpha)

qomd(p, alpha)

romd(n, alpha)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{domd} gives the density, \code{pomd} gives the distribution function, \code{qomd} gives the quantile function and \code{romd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for on the muth distribution distribution with parameters shape and scale.
}
\details{
On the muth distribution distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
domd(c(1:5),.2)
domd(2,c(.2,.3,.4))
domd(c(1:5),c(.2,.3,.4))
pomd(c(1:5),.2)
pomd(c(-1,-2,-3),.2)
pomd(2,c(.2,.3,.4))
pomd(c(1:5),c(.2,.3,.4))
pomd(-4,c(.2,.3,.4))
pomd(.4,c(.2,.3,.4))
qomd(.8,1)
qomd(.9,5)
romd(10,1)
}
\references{
Jodrá, P., Jiménez-Gamero, M. D. ve Alba-Fernández, M. V., 2015,
\emph{On the Muth distribution, Mathematical Modelling and Analysis}, 20 (3), 291-310.
}
