% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dld.R
\name{pdld}
\alias{pdld}
\title{The discrete Lindley distribution}
\usage{
pdld(x, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{ddld} gives the density, \code{pdld} gives the distribution function, \code{qdld} gives the quantile function and \code{rdld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for discrete Lindley distribution with parameters shape and scale.
}
\details{
The discrete Lindley distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
pdld(2,1)
pdld(1,c(2:4))
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
