% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pldd.R
\name{dpldd}
\alias{dpldd}
\alias{ppldd}
\alias{qpldd}
\alias{rpldd}
\title{A Power Log Dagum Distribution}
\usage{
dpldd(x, alpha, beta, theta)

ppldd(x, alpha, beta, theta)

qpldd(p, alpha, beta, theta)

rpldd(n, alpha, beta, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{beta}{shape2}

\item{theta}{shape3}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dpldd} gives the density, \code{ppldd} gives the distribution function, \code{qpldd} gives the quantile function and \code{rpldd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a Power Log Dagum distribution with parameters shape and scale.
}
\details{
A Power Log Dagum Distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dpldd(1,2,3,4)
dpldd(c(1:5),2,3,4)
dpldd(c(1:5),c(2:6),3,4)
dpldd(c(1:5),c(2:6),c(3:5),4)
dpldd(c(1:5),c(2:6),c(3:5),c(4:6))
dpldd(1,2,3,c(4:6))
ppldd(-1,2,3,4)
ppldd(c(1:3),2,3,4)
ppldd(1,c(1:3),3,4)
ppldd(1,2,c(1:4),4)
ppldd(1,2,3,c(1:4))
ppldd(c(-1:3),c(1:3),3,4)
ppldd(c(1:5),c(2:6),3,4)
ppldd(c(1:5),c(2:6),c(3:6),1)
ppldd(c(1:5),c(2:6),c(3:6),c(1:4))
qpldd(c(.7,.8),1,2,3)
qpldd(c(.7,.8),c(1,2),2,3)
qpldd(.8,1,2,3)
qpldd(c(.7,.8),c(1,2),c(2:4),3)
qpldd(.7,1,4,3)
rpldd(100,1,2,1)
}
\references{
Bakouch, H. S., Khan, M. N., Hussain, T. ve Chesneau, C., 2019,
\emph{A power log-Dagum distribution: estimation and applications}, Journal of Applied Statistics, 46 (5), 874-892.
}
