% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rld.R
\name{prld}
\alias{prld}
\title{Inferences on stress–strength reliability based on ranked set sampling data incase of Lindley distribution}
\usage{
prld(x, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{drld} gives the density, \code{prld} gives the distribution function, \code{qrld} gives the quantile function and \code{rrld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for Inferences on stress–strength reliability based on ranked set sampling data incase of Lindley distribution with parameters shape and scale.
}
\details{
Inferences on stress–strength reliability based on ranked set sampling data incase of Lindley distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
prld(c(1:5),2)
prld(2,c(2:4))
prld(c(1:5),c(2:5))
prld(-1,c(2:4))
prld(c(-1,-2),c(2:5))
prld(c(1:5),-2)
prld(-2,c(2:4))
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
