% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugd.R
\name{pugd}
\alias{pugd}
\title{Uniform-Geometric distribution}
\usage{
pugd(x, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{dugd} gives the density, \code{pugd} gives the distribution function, \code{qugd} gives the quantile function and \code{rugd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Uniform-Geometric distribution with parameters shape and scale.
}
\details{
The Uniform-Geometric distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
pugd(c(1:5),.2)
pugd(2,c(.2,.3,.4))
pugd(c(1:5),c(.2,.3,.4,.5))
pugd(-2,c(.2,.3,.4))
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
