% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uigd.R
\name{puigd}
\alias{puigd}
\title{The unit inverse Gaussian distribution A new alternative to two parameter distributions on the unit interval}
\usage{
puigd(x, mu, lambda)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{mu}{a shape parameter.}

\item{lambda}{shape2}
}
\value{
\code{duigd} gives the density, \code{puigd} gives the distribution function, \code{quigd} gives the quantile function and \code{ruigd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the unit inverse Gaussian distribution A new alternative to two parameter distribution with parameters shape and scale.
}
\details{
The unit inverse Gaussian distribution A new alternative to two parameter distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
puigd(-1,2,3)
puigd(c(1:5),2,3)
puigd(c(1:5),c(2:6),3)
puigd(c(-7),c(2:6),3)
puigd(c(1:5),c(2:6),c(3:6))
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
