% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{qemd}
\alias{qemd}
\title{Estimation in Maxwell distribution with randomly censored data}
\usage{
qemd(p, theta)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{demd} gives the density, \code{pemd} gives the distribution function, \code{qemd} gives the quantile function and \code{remd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for Estimation in Maxwell distribution with parameters shape and scale.
}
\details{
Estimation in Maxwell distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qemd(c(.7,.8),2)
qemd(.4,5)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
