% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epkd.R
\name{qepkd}
\alias{qepkd}
\title{Estimation procedures for kumaraswamy distribution parameters under adaptive type-II hybrid progressive censoring}
\usage{
qepkd(p, lambda, alpha)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{lambda}{a shape1 parameter.}

\item{alpha}{a shape parameter.}
}
\value{
\code{depkd} gives the density, \code{pepkd} gives the distribution function, \code{qepkd} gives the quantile function and \code{repkd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for Estimation procedures for kumaraswamy distribution with parameters shape and scale.
}
\details{
...Estimation procedures for kumaraswamy distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qepkd(.8,1,2)
qepkd(.4,2,5)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
