% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld.R
\name{qgld}
\alias{qgld}
\title{The gamma-Lomax distribution}
\usage{
qgld(p, a, alpha, beta)
}
\arguments{
\item{a}{a shape parameter.}

\item{alpha}{a shape1 parameter.}

\item{beta}{a shape2 parameter.}

\item{x}{new numeric vector of data values.}
}
\value{
\code{dgld} gives the density, \code{pgld} gives the distribution function, \code{qgld} gives the quantile function and \code{rgld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the gamma-Lomax distribution with parameters shape and scale.
}
\details{
The gamma-Lomax distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qgld(c(.7,.8),1,2,3)
qgld(c(.7,.8),c(1,2),2,3)
qgld(.8,2,2,3)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
