% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndd.R
\name{qndd}
\alias{qndd}
\title{A new discrete distribution}
\usage{
qndd(p, theta)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{dndd} gives the density, \code{pndd} gives the distribution function, \code{qndd} gives the quantile function and \code{rndd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a new discrete distribution with parameters shape and scale.
}
\details{
A new discrete distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qndd(.5,.2)
qndd(.8,1)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
