% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nld.R
\name{qnld}
\alias{qnld}
\title{A new lifetime distribution}
\usage{
qnld(p, lambda, beta)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{lambda}{a shape parameter.}

\item{beta}{a shape1 parameter.}
}
\value{
\code{dnld} gives the density, \code{pnld} gives the distribution function, \code{qnld} gives the quantile function and \code{rnld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a new lifetime distribution with parameters shape and scale.
}
\details{
A new lifetime distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qnld(.8,2,3)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
