% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omd.R
\name{qomd}
\alias{qomd}
\title{on the muth distribution}
\usage{
qomd(p, alpha)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}
}
\value{
\code{domd} gives the density, \code{pomd} gives the distribution function, \code{qomd} gives the quantile function and \code{romd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for on the muth distribution distribution with parameters shape and scale.
}
\details{
On the muth distribution distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qomd(.8,1)
qomd(.9,5)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
