% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.dist.R
\name{qTPMD}
\alias{qTPMD}
\title{The Power Muth Distribution}
\usage{
qTPMD(p, beta, alpha)
}
\arguments{
\item{p}{vector of probabilities.}

\item{beta}{a shape parameter.}

\item{alpha}{a scale parameter.}
}
\value{
\link{dTPMD} gives the density, \link{pTPMD} gives the distribution function, \link{qTPMD} gives the quantile function and \link{rTPMD} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Power Muth distribution with parameters shape and scale.
}
\details{
The Power Muth Distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{"(alpha/(beta^alpha))*(x^(alpha-1))*(exp((x/beta)^alpha)-1)*exp(((x/beta)^alpha)-(exp((x/beta)^alpha)-1))"}}{{\eqn{\\ f\left( x\right) =\frac{\alpha }{\beta ^{\alpha }}x^{\alpha -1}\left\{ \exp\left( \left( \frac{x}{\beta }\right) ^{\alpha }-1\right) \right\} \left\{\exp \left( \left( \frac{x}{\beta }\right) ^{\alpha }-\left( \exp \left(\frac{x}{\beta }\right) ^{\alpha }-1\right) \right) \right\}}}}.
}
\examples{
qTPMD(.5,5,2)
qTPMD(.3,2,5)
}
\references{
Jodra, P., Gomez, H. W., Jimenez-Gamero, M. D., & Alba-Fernandez, M. V. (2017).
\emph{The power Muth distribution} . Mathematical Modelling and Analysis, 22(2), 186-201.
}
