% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tprd.R
\name{qtprd}
\alias{qtprd}
\title{Two-Parameter Rayleigh Distribution}
\usage{
qtprd(p, lambda, mu)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{lambda}{a shape1 parameter.}

\item{mu}{a shape parameter.}
}
\value{
\code{dtprd} gives the density, \code{ptprd} gives the distribution function, \code{qtprd} gives the quantile function and \code{rtprd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Two-Parameter Rayleigh  distribution with parameters shape and scale.
}
\details{
The Two-Parameter Rayleigh distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qtprd(.5,1,2)
qtprd(.3,4,2)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
