% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgd.R
\name{qwbgd}
\alias{qwbgd}
\title{Weighted Bivariate Geometric Distribution}
\usage{
qwbgd(p, alpha, lambda)
}
\arguments{
\item{p}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{lambda}{a shape parameter.}
}
\value{
\code{dwbgd} gives the density, \code{pwbgd} gives the distribution function, \code{qwbgd} gives the quantile function and \code{rwbgd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Weighted Bivariate Geometric distribution with parameters shape and scale.
}
\details{
The Weighted Bivariate Geometric distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
qwbgd(c(.657),.5,1)
qwbgd(.98,.2,2)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
