% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rld.R
\name{drld}
\alias{drld}
\alias{prld}
\alias{qrld}
\alias{rrld}
\title{Inferences on stress–strength reliability based on ranked set sampling data incase of Lindley distribution}
\usage{
drld(x, theta)

prld(x, theta)

qrld(p, theta)

rrld(n, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{drld} gives the density, \code{prld} gives the distribution function, \code{qrld} gives the quantile function and \code{rrld} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for Inferences on stress–strength reliability based on ranked set sampling data incase of Lindley distribution with parameters shape and scale.
}
\details{
Inferences on stress–strength reliability based on ranked set sampling data incase of Lindley distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
drld(c(1:5),2)
drld(2,c(2:4))
drld(c(1:5),c(2:5))
drld(c(1:5),-2)
drld(-2,c(2:4))
prld(c(1:5),2)
prld(2,c(2:4))
prld(c(1:5),c(2:5))
prld(-1,c(2:4))
prld(c(-1,-2),c(2:5))
prld(c(1:5),-2)
prld(-2,c(2:4))
qrld(.8,1)
qrld(.9,5)
rrld(10,1)
}
\references{
Akgül, F. G., Acıtaş, Ş. ve Şenoğlu, B., 2018,
\emph{Inferences on stress–strength reliability based on ranked set sampling data in case of Lindley distribution}, Journal of statistical computation and simulation, 88 (15), 3018-3032.
}
