% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noPDD.R
\name{rnoPDD}
\alias{rnoPDD}
\title{A new one parameter discrete distribution and its applications}
\usage{
rnoPDD(n, theta)
}
\arguments{
\item{n}{new numeric vector of data values.}

\item{theta}{a shape parameter.}
}
\value{
\code{dnoPDD} gives the density, \code{pnoPDD} gives the distribution function, \code{qnoPDD} gives the quantile function and \code{rnoPDD} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a new one parameter discrete distribution with parameters shape and scale.
}
\details{
A new one parameter discrete distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
rnoPDD(10,1)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
