% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrd.R
\name{rsgrd}
\alias{rsgrd}
\title{Slashed generalized Rayleigh distribution}
\usage{
rsgrd(n, theta, alpha)
}
\arguments{
\item{n}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{alpha}{a shape1 parameter.}
}
\value{
\code{dsgrd} gives the density, \code{psgrd} gives the distribution function, \code{qsgrd} gives the quantile function and \code{rsgrd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Slashed generalized Rayleigh distribution with parameters shape and scale.
}
\details{
The Slashed generalized Rayleigh distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
rsgrd(10,1,2)
}
\references{
Gómez-Déniz, E. ve Calderín-Ojeda, E., 2011,
\emph{The discrete Lindley distribution: properties and applications}.Journal of statistical computation and simulation, 81 (11), 1405-1416.
}
