% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrd.R
\name{dsgrd}
\alias{dsgrd}
\alias{psgrd}
\alias{qsgrd}
\alias{rsgrd}
\title{Slashed generalized Rayleigh distribution}
\usage{
dsgrd(x, theta, alpha)

psgrd(x, theta, alpha)

qsgrd(p, theta, alpha)

rsgrd(n, theta, alpha)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{alpha}{a shape1 parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dsgrd} gives the density, \code{psgrd} gives the distribution function, \code{qsgrd} gives the quantile function and \code{rsgrd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Slashed generalized Rayleigh distribution with parameters shape and scale.
}
\details{
The Slashed generalized Rayleigh distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dsgrd(1,2,3)
dsgrd(c(1:5),2,3)
dsgrd(c(1:5),c(2:6),3)
dsgrd(c(1:5),c(2:6),c(3:5))
psgrd(1,2,3)
psgrd(c(1:5),2,3)
psgrd(c(1:5),c(2:6),3)
psgrd(c(1:5),c(2:6),c(3:5))
psgrd(-5,c(2:6),c(3:5))
psgrd(.1,1,2)
qsgrd(c(.7,.8),2,3)
qsgrd(c(.7,.8),c(1,2),2)
rsgrd(10,1,2)
}
\references{
Iriarte, Y. A., Vilca, F., Varela, H. ve Gómez, H. W., 2017,
\emph{Slashed generalized Rayleigh distribution}, Communications in Statistics-Theory and Methods, 46 (10), 4686-4699.
}
