% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tprd.R
\name{dtprd}
\alias{dtprd}
\alias{ptprd}
\alias{qtprd}
\alias{rtprd}
\title{Two-Parameter Rayleigh Distribution}
\usage{
dtprd(x, lambda, mu)

ptprd(x, lambda, mu)

qtprd(p, lambda, mu)

rtprd(n, lambda, mu)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{lambda}{a shape1 parameter.}

\item{mu}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dtprd} gives the density, \code{ptprd} gives the distribution function, \code{qtprd} gives the quantile function and \code{rtprd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Two-Parameter Rayleigh  distribution with parameters shape and scale.
}
\details{
The Two-Parameter Rayleigh distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dtprd(c(5),c(4),5)
dtprd(c(9),c(4:8),5)
dtprd(c(5),c(3,4),c(4:9))
dtprd(c(5),c(4),-5)
ptprd(c(4:5),c(4),3)
ptprd(c(6),c(4:5),4)
ptprd(4,1,c(1:3))
ptprd(-5,c(4:8),4)
qtprd(.5,1,2)
qtprd(.3,4,2)
rtprd(100,1,2)
}
\references{
Dey, S., Dey, T. ve Kundu, D., 2014,
\emph{Two-parameter Rayleigh distribution: different methods of estimation}, American Journal of Mathematical and Management Sciences, 33 (1), 55-74.
}
