% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugd.R
\name{ugd}
\alias{ugd}
\alias{dugd}
\alias{pugd}
\alias{qugd}
\alias{rugd}
\title{Uniform-Geometric distribution}
\usage{
dugd(x, theta)

pugd(x, theta)

qugd(p, theta)

rugd(n, theta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{theta}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dugd} gives the density, \code{pugd} gives the distribution function, \code{qugd} gives the quantile function and \code{rugd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Uniform-Geometric distribution with parameters shape and scale.
}
\details{
The Uniform-Geometric distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dugd(c(1:5),.2)
dugd(2,c(.2,.4))
dugd(c(1:5),c(.2,.3,.4,.5))
pugd(c(1:5),.2)
pugd(2,c(.2,.3,.4))
pugd(c(1:5),c(.2,.3,.4,.5))
pugd(-2,c(.2,.3,.4))
qugd(0.6,.1)
rugd(100,.1)
}
\references{
Akdoğan, Y., Kuş, C., Asgharzadeh, A., Kınacı, İ., & Sharafi, F. (2016).
\emph{Uniform-geometric distribution}. Journal of Statistical Computation and Simulation, 86(9), 1754-1770.
}
