% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uigd.R
\name{uigd}
\alias{uigd}
\alias{duigd}
\alias{puigd}
\alias{quigd}
\alias{ruigd}
\title{The unit inverse Gaussian distribution A new alternative to two parameter distributions on the unit interval}
\usage{
duigd(x, mu, lambda)

puigd(x, mu, lambda)

quigd(p, mu, lambda)

ruigd(n, mu, lambda)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{mu}{a shape parameter.}

\item{lambda}{shape2}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{duigd} gives the density, \code{puigd} gives the distribution function, \code{quigd} gives the quantile function and \code{ruigd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the unit inverse Gaussian distribution A new alternative to two parameter distribution with parameters shape and scale.
}
\details{
The unit inverse Gaussian distribution A new alternative to two parameter distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
duigd(1,2,3)
duigd(c(1:5),2,3)
duigd(c(1:5),c(2:6),3)
duigd(c(1:5),c(2:6),c(3:6))
puigd(-1,2,3)
puigd(c(1:5),2,3)
puigd(c(1:5),c(2:6),3)
puigd(c(-7),c(2:6),3)
puigd(c(1:5),c(2:6),c(3:6))
quigd(.1,1,1)
ruigd(100,1,1)
}
\references{
Ghitany, M., Mazucheli, J., Menezes, A. ve Alqallaf, F., 2019,
\emph{The unit-inverse Gaussian distribution: A new alternative to two-parameter distributions on the unit interval}, Communications in Statistics-Theory and Methods, 48 (14), 3423-3438.
}
