% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.dist.R
\name{dTMPD}
\alias{dTMPD}
\title{The Power Muth Distribution}
\usage{
dTMPD(x, beta, alpha)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{beta}{a shape parameter.}

\item{alpha}{a scale parameter.}
}
\value{
\code{\link[=dTMPD]{dTMPD()}} returns the density.
}
\description{
\code{\link[=dTMPD]{dTMPD()}} performs a density of The Power Muth Distribution.
}
\details{
The Power Muth Distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
}
\examples{
dTMPD(1,2,3)
dTMPD(2,3,4)
dTMPD(c(1:5),2,3)
dTMPD(c(1:5),c(2:6),3)
dTMPD(c(1:5),c(2:6),c(3:6))
}
\references{
Jodra, P., Gomez, H. W., Jimenez-Gamero, M. D., & Alba-Fernandez, M. V. (2017).
\emph{The power Muth distribution} . Mathematical Modelling and Analysis, 22(2), 186-201.
}
