% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nld.R
\name{nld}
\alias{nld}
\alias{dnld}
\alias{pnld}
\alias{qnld}
\alias{rnld}
\title{A new lifetime distribution}
\usage{
dnld(x, lambda, beta, log = FALSE)

pnld(x, lambda, beta, lower.tail = TRUE, log.p = FALSE)

qnld(p, lambda, beta, lower.tail = TRUE)

rnld(n, lambda, beta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda, beta}{are parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P\left[ X\leq x\right]}, otherwise,\eqn{P\left[ X>x\right] }.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}
}
\value{
\code{dnld} gives the density, \code{pnld} gives the distribution
function, \code{qnld} gives the quantile function and \code{rnld} generates
random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
a new lifetime distribution parameters.
}
\details{
A new lifetime distribution with parameters are \eqn{\lambda}, \eqn{\beta},
has density given by
\deqn{f\left( x\right) =\frac{\lambda \beta }
{\left( 1-e^{-\lambda }\right) } e^{-\lambda -\beta x+\lambda e^{-\beta x}},}
where
\deqn{x>\mathbb{R}_{+},~\beta ,\lambda \in \mathbb{R}_{+}.}
}
\examples{
library(new.dist)
dnld(1, lambda=2, beta=3)
pnld(1,lambda=2,beta=3)
qnld(.8,lambda=2,beta=3)
rnld(10,lambda=2,beta=3)
}
\references{
Kuş, C., 2007,
\emph{A new lifetime distribution}, Computational Statistics & Data Analysis,
51 (9), 4497-4509.
}
