% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sod.R
\name{sod}
\alias{sod}
\alias{dsod}
\alias{psod}
\alias{qsod}
\alias{rsod}
\title{On parameter estimation of the standard omega distribution}
\usage{
dsod(x, alpha, beta)

psod(x, alpha, beta)

qsod(p, alpha, beta)

rsod(n, alpha, beta)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{beta}{a shape1 parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dsod} gives the density, \code{psod} gives the distribution function, \code{qsod} gives the quantile function and \code{rsod} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for On parameter estimation of the standard omega distribution with parameters shape and scale.
}
\details{
On parameter estimation of the standard omega distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
library(new.dist)
dsod(c(0.4),c(4),5)
dsod(c(0.4),c(4:8),5)
dsod(c(0.3,0.4),c(3,4),c(4:9))
psod(c(0.4),c(4),5)
psod(c(0.3,0.4),c(3,4),c(4,5))
psod(c(-2),c(3,4),c(4,5))
qsod(.8,2,3)
rsod(10,1,2)
}
\references{
Birbiçer, İ. ve Genç, A. İ., 2022,
\emph{On parameter estimation of the standard omega distribution}. Journal of Applied Statistics, 1-17.
}
