% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgd.R
\name{dwbgd}
\alias{dwbgd}
\alias{pwbgd}
\alias{qwbgd}
\alias{rwbgd}
\title{Weighted Bivariate Geometric Distribution}
\usage{
dwbgd(x, alpha, lambda)

pwbgd(x, alpha, lambda)

qwbgd(p, alpha, lambda)

rwbgd(n, alpha, lambda)
}
\arguments{
\item{x}{new numeric vector of data values.}

\item{alpha}{a shape parameter.}

\item{lambda}{a shape parameter.}

\item{p}{new numeric vector of data values.}

\item{n}{new numeric vector of data values.}
}
\value{
\code{dwbgd} gives the density, \code{pwbgd} gives the distribution function, \code{qwbgd} gives the quantile function and \code{rwbgd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the Weighted Bivariate Geometric distribution with parameters shape and scale.
}
\details{
The Weighted Bivariate Geometric distribution with shape parameter \ifelse{html}{\out{alpha}}{\eqn{\alpha}} and shape1 parameter \ifelse{html}{\out{beta}}{\eqn{\beta}}and scale parameter \ifelse{html}{\out{beta}}{\eqn{\beta}} has density
\ifelse{html}{\out{T<sub>n</sub><sup>(&#8467)</sup>}}{\eqn{T_n^{(\ell)}}}.
}
\examples{
dwbgd(1,.2,3)
dwbgd(c(1:5),.2,3)
dwbgd(c(1:5),c(.2,.3,.4),3)
dwbgd(c(1:5),c(.2,.3,.4),c(3:5))
pwbgd(1,.2,3)
pwbgd(c(1:5),.2,3)
pwbgd(c(1:5),c(.2,.6),3)
pwbgd(c(1:5),c(.2,.6),c(3:5))
pwbgd(-5,c(.2,.6),c(3:5))
qwbgd(c(.657),.5,1)
qwbgd(.98,.2,2)
qwbgd(c(.657),.5,1)
qwbgd(.98,.2,2)
}
\references{
Najarzadegan, H., Alamatsaz, M. H., Kazemi, I. ve Kundu, D., 2020,
\emph{Weighted bivariate geometric distribution: Simulation and estimation}, Communications in Statistics-Simulation and Computation, 49 (9), 2419-2443.
}
